Organize all classes into a folder & Update require_relatives
